[Feat] 몬스터 이동속도 로직 수정 및 보스5 프리팹 컴포넌트  부착

+++ Assets/Scripts/YTH/Monster/MonsterModel.cs
@@ -16,6 +16,6 @@ public class MonsterModel : MonoBehaviour
     [SerializeField] int _curHp;
     public int CurHp { get => _curHp; set => _curHp = value; }
 
-    private int _moveSpeed;
-    public int MoveSpeed { get => _moveSpeed; set => _moveSpeed = value; }
+    private float _moveSpeed;
+    public float MoveSpeed { get => _moveSpeed; set => _moveSpeed = value; }
 }
